Final updates

+++ Domain/PlaneMinimal.cs
@@ -4,7 +4,7 @@ public class PlaneMinimal
 {
     public long Time {get; set;} = 0;
     public string HexValue {get; set;} = string.Empty;
-    public int Latitude {get; set;} = 0;
-    public int Longitude {get; set;} = 0;
-    public int Altitude {get; set;} = 0;
+    public float Latitude {get; set;} = 0;
+    public float Longitude {get; set;} = 0;
+    public float Altitude {get; set;} = 0;
 }
